flask_app/requirements.txt
flask
requests
pillow
matplotlib
openpyxl
scipy
pandas
numpy
opencv-python
torch
torchvision
matplotlib
tqdm
scikit-learn
boto3
python-dotenv==1.0.1